Correccion tipo de id foro

+++ foro/src/main/java/uni/fis/foro/controller/ForoCategoriaController.java
@@ -30,7 +30,7 @@ public class ForoCategoriaController {
     }
 
     @GetMapping("/categoria/{idCategoria}")
-    public ResponseEntity<List<ForoCategoriaEntity>> listarPorCategoria(@PathVariable Long idCategoria) {
+    public ResponseEntity<List<ForoCategoriaEntity>> listarPorCategoria(@PathVariable Integer idCategoria) {
         return ResponseEntity.ok(foroCategoriaService.listarPorCategoria(idCategoria));
     }
 }
